refactor(roles): split store responsibilities
